moved login from pages to users
